Refactor feature popover
* Reuse site components in popup and popover
